+++ src/Http/Controllers/FormSubmissionController.php
@@ -2,14 +2,19 @@
 
 namespace Marble\Admin\Http\Controllers;
 
+use Illuminate\Foundation\Auth\Access\AuthorizesRequests;
 use Illuminate\Routing\Controller;
 use Marble\Admin\Models\FormSubmission;
 use Marble\Admin\Models\Item;
 
 class FormSubmissionController extends Controller
 {
+    use AuthorizesRequests;
+
     public function index(Item $item)
     {
+        $this->authorize('update', $item);
+
         $submissions = FormSubmission::where('item_id', $item->id)
             ->orderByDesc('created_at')
             ->paginate(30);
@@ -19,6 +24,8 @@ class FormSubmissionController extends Controller
 
     public function show(Item $item, FormSubmission $submission)
     {
+        $this->authorize('update', $item);
+
         $submission->update(['read' => true]);
 
         return view('marble::form.show', compact('item', 'submission'));
@@ -26,6 +33,8 @@ class FormSubmissionController extends Controller
 
     public function markRead(Item $item, FormSubmission $submission)
     {
+        $this->authorize('update', $item);
+
         $submission->update(['read' => true]);
 
         return redirect()->route('marble.item.edit', $item);
@@ -33,6 +42,8 @@ class FormSubmissionController extends Controller
 
     public function destroy(Item $item, FormSubmission $submission)
     {
+        $this->authorize('update', $item);
+
         $submission->delete();
 
         return redirect()->route('marble.form.index', $item);
